backend/requirements.txt
Django==5.2.8
Djangorestframework==3.16.1